Extract linter commands into their own module under tmt.cli
Part of the ongoing effort to shorten tmt.cli code.
