+++ tmt/__main__.py
@@ -7,6 +7,7 @@ def import_cli_commands() -> None:
     # TODO: some kind of `import tmt.cli.*` would be nice
     import tmt.cli._root  # type: ignore[reportUnusedImport,unused-ignore]
     import tmt.cli.init  # noqa: F401,I001,RUF100
+    import tmt.cli.lint  # noqa: F401,I001,RUF100
     import tmt.cli.status  # noqa: F401,I001,RUF100
     import tmt.cli.trying  # noqa: F401,I001,RUF100
